Separate result structure & entities' structure
`die-values` was embedded in the `dice-ast` before `dice-ast` is even evaluated.
To prevent this, we make a distinction between evaluation result structure & the structure which represents the entities.

This commit also implements the reroll functionality. You will also see that instead of passing a literal set, I now pass a list of `Die`s to the various operators(`keep-op`, `remove-op`, `reroll-op`.
The above will help in adding & modifying the `previous-values` & `discarded` keys to `Die` in the future.