Handle when there expression results in infinite rerolls
Imgaine an expression `4d1rrl4` which says that keep rerolling until there is no value left in the set of lowest 4 values in `4d1`.

Since the dice has only 1 face, everytime we reroll we get 1 & that 1 is one of the values in the original lowest 4 values' set.
This means we will never get a value which won't be in the original lowest 4 values' set & thus lead to infinite scroll.

This change handles the above scenario.