Record `previous-values` or `discarded` on an operation
When rerolling, we now record values which do not pass our criteria.
Rerolling & recording `previous-values` works on all selectors.

When `keep`ing or `delete`ing a `Die`, we set `discarded` true to the `Die`s we are removing from the set.
`keep` & `delete` works on all selectors except `lowest` & `highest` for now.